# server/api/api.yaml
...
openapi: 3.0.0
info:
  version: 0.6.0
  title: Playbooks API
  contact:
    name: ServiceExchange
    url: https://service-exchange.com/
    email: support@service-exchange.com
servers:
  - url: http://localhost:8065/plugins/playbooks/api/v0
paths:
...
  /runs/{id}:
    get:
      summary: Get a playbook run
      operationId: getPlaybookRun
      security:
        - BearerAuth: []
      tags:
        - PlaybookRuns
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the playbook run to retrieve.
          schema:
            type: string
            example: mx3xyzdojfgyfdx8sc8of1gdme
      x-codeSamples:
        - lang: curl
          source: |
            curl -X GET 'http://localhost:8065/plugins/playbooks/api/v0/runs/mx3xyzdojfgyfdx8sc8of1gdme' \
            -H 'Authorization: Bearer 9g64ig7q9pds8yjz8rsgd6e36e'
      responses:
        200:
          description: Playbook run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookRun"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
    patch:
      summary: Update a playbook run
      operationId: updatePlaybookRun
      security:
        - BearerAuth: []
      tags:
        - PlaybookRuns
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the playbook run to retrieve.
          schema:
            type: string
            example: mx3xyzdojfgyfdx8sc8of1gdme
      requestBody:
        description: Playbook run update payload.
        content:
          application/json:
            schema:
              type: object
              properties:
                active_stage:
                  type: integer
                  description: Zero-based index of the stage that will be made active.
                  example: 2
      x-codeSamples:
        - lang: curl
          source: |
            curl -X PATCH 'http://localhost:8065/plugins/playbooks/api/v0/runs/mx3xyzdojfgyfdx8sc8of1gdme' \
            -H 'Authorization: Bearer 9g64ig7q9pds8yjz8rsgd6e36e'\
            -H 'Content-Type: application/json'\
            -d '{"active_stage": 2}'
      responses:
        200:
          description: Playbook run successfully updated.
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    400:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: The request is malformed.
    403:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: Access to the resource is forbidden for this user.
...
    500:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: There was an internal error in the server.
  schemas:
    PlaybookRun:
      type: object
      properties:
        id:
          type: string
          description: A unique, 26 characters long, alphanumeric identifier for the playbook run.
          example: mx3xyzdojfgyfdx8sc8of1gdme
        name:
          type: string
          description: The name of the playbook run.
          example: Server down in EU cluster
        description:
          type: string
          description: The description of the playbook run.
          example: There is one server in the EU cluster that is not responding since April 12.
        is_active:
          type: boolean
          description: True if the playbook run is ongoing; false if the playbook run is ended.
          example: false
        owner_user_id:
          type: string
          description: The identifier of the user that is commanding the playbook run.
          example: bqnbdf8uc0a8yz4i39qrpgkvtg
        team_id:
          type: string
          description: The identifier of the team where the playbook run's channel is in.
          example: 61ji2mpflefup3cnuif80r5rde
        channel_id:
          type: string
          description: The identifier of the playbook run's channel.
          example: hwrmiyzj3kadcilh3ukfcnsbt6
        create_at:
          type: integer
          format: int64
          description: The playbook run creation timestamp, formatted as the number of milliseconds since the Unix epoch.
          example: 1606807976289
        end_at:
          type: integer
          format: int64
          description: The playbook run finish timestamp, formatted as the number of milliseconds since the Unix epoch. It equals 0 if the playbook run is not finished.
          example: 0
        delete_at:
          type: integer
          format: int64
          description: The playbook run deletion timestamp, formatted as the number of milliseconds since the Unix epoch. It equals 0 if the playbook run is not deleted.
          example: 0
        active_stage:
          type: integer
          format: int32
          description: Zero-based index of the currently active stage.
          example: 1
        active_stage_title:
          type: string
          description: The title of the currently active stage.
          example: Triage issue
        post_id:
          type: string
          description: If the playbook run was created from a post, this field contains the identifier of such post. If not, this field is empty.
          example: b2ntfcrl4ujivl456ab4b3aago
        playbook_id:
          type: string
          description: The identifier of the playbook with from which this playbook run was created.
          example: 0y4a0ntte97cxvfont8y84wa7x
        checklists:
          type: array
          items:
            $ref: "#/components/schemas/Checklist"
...
    Checklist:
      type: object
      properties:
        id:
          type: string
          description: A unique, 26 characters long, alphanumeric identifier for the checklist.
          example: 6f6nsgxzoq84fqh1dnlyivgafd
        title:
          type: string
          description: The title of the checklist.
          example: Triage issue
        items:
          type: array
          description: The list of tasks to do.
          items:
            $ref: "#/components/schemas/ChecklistItem"
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
          description: A unique, 26 characters long, alphanumeric identifier for the checklist item.
          example: 6f6nsgxzoq84fqh1dnlyivgafd
        title:
          type: string
          description: The title of the checklist item.
          example: Gather information from customer.
        state:
          type: string
          enum:
            - ""
            - in_progress
            - closed
          description: The state of the checklist item. An empty string means that the item is not done.
          example: closed
        state_modified:
          type: integer
          format: int64
          description: The timestamp for the latest modification of the item's state, formatted as the number of milliseconds since the Unix epoch. It equals 0 if the item was never modified.
          example: 1607774621321
        assignee_id:
          type: string
          description: The identifier of the user that has been assigned to complete this item. If the item has no assignee, this is an empty string.
          example: pisdatkjtdlkdhht2v4inxuzx1
        assignee_modified:
          type: integer
          format: int64
          description: The timestamp for the latest modification of the item's assignee, formatted as the number of milliseconds since the Unix epoch. It equals 0 if the item never got an assignee.
          example: 1608897821125
        command:
          type: string
          description: The slash command associated with this item. If the item has no slash command associated, this is an empty string
          example: /opsgenie on-call
        command_last_run:
          type: integer
          format: int64
          description: The timestamp for the latest execution of the item's command, formatted as the number of milliseconds since the Unix epoch. It equals 0 if the command was never executed.
          example: 1608552221019
        description:
          type: string
          description: A detailed description of the checklist item, formatted with Markdown.
          example: Ask the customer for more information in [Zendesk](https://www.zendesk.com/).
    Error:
      type: object
      required:
        - error
        - details
      properties:
        error:
          type: string
          description: A message with the error description.
          example: Error retrieving the resource.
        details:
          type: string
          description: Further details on where and why this error happened.
          example: Specific details about the error, depending on the case.
...